+++ test/src/data/mocks/hive_mock.dart
@@ -3,26 +3,62 @@ import 'dart:typed_data';
 
 import 'package:hive/hive.dart';
 import 'package:mockito/mockito.dart';
+import 'package:string_manager_flutter/src/data/constants/constants.dart';
 import 'package:string_manager_flutter/src/data/models/string_resource.dart';
 
+import '../constants/constants.dart';
+
 class MockBox<E> extends Mock implements Box<E> {
+  final bool returnDefaults;
+
+  MockBox(this.returnDefaults) {
+    if (returnDefaults) {
+      language = 'en';
+      resource = StringResource(resource: {
+        iAmABoy: iAmABoy,
+        itIsPlenty: itIsPlenty,
+      });
+    }
+  }
+
   StringResource resource = StringResource();
+  String? language;
+
+  bool open = false;
+
+  @override
+  bool get isOpen => open;
 
   @override
   E? get(key, {E? defaultValue}) {
-    return resource as E;
+    if (key == Constants.languageStorageKey) {
+      return language as E?;
+    } else {
+      return resource as E;
+    }
   }
 
   @override
   Future<void> put(key, E value) async {
+    if (value is String) {
+      language = value;
+      return;
+    }
     resource = value as StringResource;
   }
 
-
-
+  @override
+  Future<void> close() async {
+    open = false;
+    dev.log('close called');
+  }
 }
 
 class HiveMock extends Mock implements HiveInterface {
+  final bool _returnDefaults;
+
+  HiveMock({bool? returnDefaults}) : _returnDefaults = returnDefaults ?? false;
+
   @override
   void init(
     String? path, {
@@ -32,7 +68,6 @@ class HiveMock extends Mock implements HiveInterface {
     dev.log('init called');
   }
 
-
   Future<void> initFlutter([String? subDir]) async {
     dev.log('initFlutter called');
   }
@@ -41,7 +76,6 @@ class HiveMock extends Mock implements HiveInterface {
   bool isAdapterRegistered(int typeId) {
     return true;
   }
-
 
   @override
   void registerAdapter<T>(TypeAdapter<T> adapter,
@@ -66,6 +100,10 @@ class HiveMock extends Mock implements HiveInterface {
     String? collection,
     @Deprecated('Use encryptionCipher instead') List<int>? encryptionKey,
   }) {
-    return Future.value(MockBox<E>());
+    MockBox<E> mockBox = MockBox<E>(
+      _returnDefaults,
+    );
+    mockBox.open = true;
+    return Future.value(mockBox);
   }
 }