+++ CHANGELOG.md
@@ -2,3 +2,9 @@
 
 - StringManger singleton class for managing, storing and translating flutter app strings using hive
   and google_translate
+  
+
+## 0.0.2
+
+- Supports sound null safety
+- fix issue of being unable to translate stored strings when initialized in a new locale